# api/messages.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "messages.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MessagesAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/messages": {
      "get": {
        "operationId": "MessagesAPI_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uint64"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "MessagesAPI"
        ]
      },
      "delete": {
        "operationId": "MessagesAPI_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiDeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiDeleteRequest"
            }
          }
        ],
        "tags": [
          "MessagesAPI"
        ]
      },
      "post": {
        "operationId": "MessagesAPI_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateRequest"
            }
          }
        ],
        "tags": [
          "MessagesAPI"
        ]
      }
    }
  },
  "definitions": {
    "MentionMentionType": {
      "type": "string",
      "enum": [
        "Unspecified",
        "At",
        "Hash",
        "Dollar",
        "Tilda"
      ],
      "default": "Unspecified"
    },
    "apiCreateRequest": {
      "type": "object",
      "properties": {
        "threadId": {
          "type": "string",
          "format": "uint64"
        },
        "accountId": {
          "type": "string",
          "format": "uint64"
        },
        "mentions": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "uint64"
          }
        },
        "content": {
          "type": "string"
        }
      }
    },
    "apiCreateResponse": {
      "type": "object",
      "properties": {
        "message": {
          "$ref": "#/definitions/apiMessage"
        }
      }
    },
    "apiDeleteRequest": {
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "uint64"
          }
        }
      }
    },
    "apiDeleteResponse": {
      "type": "object"
    },
    "apiGetResponse": {
      "type": "object",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiMessage"
          }
        }
      }
    },
    "apiMention": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "type": {
          "$ref": "#/definitions/MentionMentionType"
        },
        "start": {
          "type": "integer",
          "format": "int64"
        },
        "end": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "apiMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "threadId": {
          "type": "string",
          "format": "uint64"
        },
        "accountId": {
          "type": "string",
          "format": "uint64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "mentions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiMention"
          }
        },
        "content": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}